Move `MessageCountsProvider` to :legacy:message` module
