+++ feature/widget/unread/src/main/kotlin/app/k9mail/feature/widget/unread/UnreadWidgetDataProvider.kt
@@ -4,13 +4,13 @@ import android.content.Context
 import android.content.Intent
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.mailstore.FolderRepository
+import app.k9mail.legacy.message.controller.MessageCountsProvider
 import app.k9mail.legacy.search.LocalSearch
 import app.k9mail.legacy.search.SearchAccount
 import app.k9mail.legacy.ui.folder.FolderNameFormatter
 import com.fsck.k9.CoreResourceProvider
 import com.fsck.k9.Preferences
 import com.fsck.k9.activity.MessageList
-import com.fsck.k9.controller.MessageCountsProvider
 import com.fsck.k9.ui.messagelist.DefaultFolderProvider
 import timber.log.Timber
