+++ legacy/core/src/main/java/com/fsck/k9/controller/KoinModule.kt
@@ -2,6 +2,7 @@ package com.fsck.k9.controller
 
 import android.content.Context
 import app.k9mail.legacy.mailstore.MessageStoreManager
+import app.k9mail.legacy.message.controller.MessageCountsProvider
 import app.k9mail.legacy.message.controller.MessagingControllerRegistry
 import com.fsck.k9.Preferences
 import com.fsck.k9.backend.BackendManager
