+++ legacy/message/build.gradle.kts
@@ -8,6 +8,7 @@ android {
 
 dependencies {
     implementation(projects.legacy.account)
+    implementation(projects.legacy.search)
 
     implementation(projects.mail.common)
     implementation(projects.backend.api)
